refactor: extract CachedJsonFile to deduplicate persistence pattern
Move CachedJsonFile class from json-store.js into its own dedicated
module (cached-json-file.js). Update config-manager and session-manager
to import directly from the new module. json-store.js re-exports the
class for backward compatibility. Also adds an invalidate() method and
fixes array default-value copying.

Closes #573
